Backend/pyproject.toml
[project]
name = "whatsapp-ai-crm"
version = "0.1.0"
description = "WhatsApp AI CRM Backend with Screaming Architecture"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "supabase>=2.3.6",
    "openai>=1.14.0",
    "google-generativeai>=0.4.1",
    "pydantic>=2.6.4",
    "pydantic-settings>=2.2.1",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "orjson>=3.9.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "coverage>=7.4.0",
]
